+++ pkg/controller/kfservice/kfservice_controller.go
@@ -153,6 +153,7 @@ func (r *ReconcileService) Reconcile(request reconcile.Request) (reconcile.Resul
 
 	reconcilers := []Reconciler{
 		knative.NewRouteReconciler(r.Client, r.scheme),
+		knative.NewConfigurationReconciler(r.Client, r.scheme, configMap),
 	}
 
 	for _, reconciler := range reconcilers {
