+++ pkg/resources/knative/route.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resources
+package knative
 
 import (
 	knservingv1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
